# Conflicts:
#	Marlin/src/Marlin.cpp
#	Marlin/src/gcode/config/M200-M205.cpp

+++ Marlin/src/gcode/config/M200-M205.cpp
@@ -61,7 +61,7 @@ void GcodeSuite::M201() {
 
   LOOP_XYZE(i) {
     if (parser.seen(axis_codes[i])) {
-      const uint8_t a = (i == E_AXIS ? E_AXIS_N(target_extruder) : i);
+      const uint8_t a = (i == E_AXIS ? uint8_t(E_AXIS_N(target_extruder)) : i);
       static constexpr uint32_t max_accel[] = MAX_ACCELERATION_LIMIT;
       planner.settings.max_acceleration_mm_per_s2[a] = constrain(parser.value_axis_units((AxisEnum)a), 10, max_accel[a]);
     }
@@ -82,7 +82,7 @@ void GcodeSuite::M203() {
 
   LOOP_XYZE(i)
     if (parser.seen(axis_codes[i])) {
-      const uint8_t a = (i == E_AXIS ? E_AXIS_N(target_extruder) : i);
+      const uint8_t a = (i == E_AXIS ? uint8_t(E_AXIS_N(target_extruder)) : i);
       static constexpr uint32_t max_feedrate[] = MAX_FEEDRATE_LIMIT;
       planner.settings.max_feedrate_mm_s[a] = constrain(parser.value_axis_units((AxisEnum)a), 10, max_feedrate[a]);
     }
